Fixed bugs with PGN formatting, added opening book support

+++ Boxfish-Lib/src/Types.h
@@ -158,7 +158,7 @@ namespace Boxfish
 
 	constexpr Square INVALID_SQUARE = { FILE_INVALID, RANK_INVALID };
 
-	enum SquareIndex : int 
+	enum SquareIndex : int8_t 
 	{
 		a1, b1, c1, d1, e1, f1, g1, h1,
 		a2, b2, c2, d2, e2, f2, g2, h2,
